moving gaussian models into sub-dir
