+++ floris/simulation/wake_velocity/gaussianModels/__init__.py
@@ -10,4 +10,4 @@
 # License for the specific language governing permissions and limitations under
 # the License.
 
-from . import gaussianModels
+
