+++ floris/simulation/wake_velocity/gaussianModels/gauss_merge.py
@@ -12,8 +12,8 @@
 
 import numpy as np
 from scipy.special import gamma
-from ...utilities import cosd, sind, tand
-from .base_velocity_deficit import VelocityDeficit
+from ....utilities import cosd, sind, tand
+from ..base_velocity_deficit import VelocityDeficit
 from .gaussian_model_ish import GaussianModel
